requirements.txt
pywin32
tk
pyautogui
keyboard
python-telegram-bot